+++ src/prefect/serialization/state.py
@@ -68,6 +68,12 @@ class ScheduledSchema(PendingSchema):
 
 
 @version("0.3.3")
+class ResumeSchema(ScheduledSchema):
+    class Meta:
+        object_class = state.Resume
+
+
+@version("0.3.3")
 class RetryingSchema(ScheduledSchema):
     class Meta:
         object_class = state.Retrying
@@ -145,6 +151,7 @@ class StateSchema(OneOfSchema):
         "CachedState": CachedStateSchema,
         "Scheduled": ScheduledSchema,
         "Retrying": RetryingSchema,
+        "Resume": ResumeSchema,
         "Running": RunningSchema,
         "Finished": FinishedSchema,
         "Success": SuccessSchema,